Moved configuration validation to spring boot main to resolve resursive bean dependency issues
