+++ oag/src/main/java/org/owasp/oag/infrastructure/OAGBeanConfiguration.java
@@ -3,7 +3,6 @@ package org.owasp.oag.infrastructure;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.owasp.oag.config.ConfigLoader;
 import org.owasp.oag.config.FileConfigLoader;
-import org.owasp.oag.config.InvalidOAGSettingsException;
 import org.owasp.oag.config.configuration.MainConfig;
 import org.owasp.oag.services.blacklist.LocalPersistentBlacklist;
 import org.owasp.oag.services.blacklist.SessionBlacklist;
@@ -13,7 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -28,26 +26,15 @@ public class OAGBeanConfiguration {
     private static final Logger log = LoggerFactory.getLogger(OAGBeanConfiguration.class);
 
     @Autowired
-    private ApplicationContext context;
-
-    @Autowired
     private GlobalClockSource clockSource;
 
     @Bean
-    public MainConfig mainConfig() throws InvalidOAGSettingsException {
+    public MainConfig mainConfig(){
 
         try {
 
             ConfigLoader loader = configLoader();
             MainConfig config = loader.loadConfiguration();
-
-            var configErrors = config.getErrors(context);
-            if (!configErrors.isEmpty()) {
-                String message = "Configuration file contains errors: " + configErrors.toString();
-                log.error(message);
-                throw new InvalidOAGSettingsException(message);
-            }
-
             return config;
 
         } catch (JsonProcessingException e) {
